added some changes

+++ src/backend/DeLong.Domain/Entities/CashRegister.cs
@@ -4,8 +4,7 @@ namespace DeLong.Domain.Entities;
 
 public class CashRegister : Auditable
 {
-    public int WarehouseId { get; set; }
+    public long WarehouseId { get; set; }
     public required Warehouse Warehouse { get; set; }
     public decimal Balance { get; set; }
 }
-
